lay ket qua hoteldetail ra API, dang lam render ra UI

+++ HotelBooking.infrastructure/Models/Review.cs
@@ -11,7 +11,7 @@ public partial class Review
 
     public int CustomerId { get; set; }
 
-    public int? Rating { get; set; }
+    public double? Rating { get; set; }
 
     public string? Comment { get; set; }
